Changed CvBuildings cached ReplaceBuildings from an array to a vector, instead of having 6000 arrays with 6000 elements cached it will now have around 300 vectors with around 1-10 elements. Improves RAM usage and performance.

+++ Sources/CvInfos.h
@@ -2284,7 +2284,7 @@ public:
 	bool isDefendAgainstUnit(int i) const;		//Exposed to Python
 
 	int getSupersedingUnit(int i) const;
-	int getNumSupersedingUnits() const;
+	short getNumSupersedingUnits() const;
 	bool isSupersedingUnit(int i) const;		//Exposed to Python
 
 	int getUnitUpgrade(int i) const;			//Exposed to Python
